open-router-002/openrouterproject-pkg/pyproject.toml
[project]
name = "openrouterproject-pkg"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "waniya shah", email = "160254821+Waniyashah@users.noreply.github.com" }
]
requires-python = ">=3.11"
dependencies = [
    "crewai>=0.121.1",
    "python-dotenv>=1.1.0",
]

[project.scripts]
openrouterproject-pkg = "openrouterproject_pkg:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
